backend/requirements.txt
sqlalchemy
alembic
passlib[bcrypt]
pyjwt